build(quality): enforce warning-clean builds with -Werror (#453) (#461)
Add -Werror to all JavaCompile tasks so any javac or Error Prone warning
of any class fails the build the cycle it appears, turning "warning-clean"
from a one-time state into an enforced invariant. Set
disableWarningsInGeneratedCode so Lombok-generated members don't trip it.

Clean-build ratchet surfaced five default-WARNING stragglers (not in the
#446-#452 promotion set), all fixed here:
- JavaDurationGetSecondsToToSeconds: getSeconds() -> toSeconds()
- LongDoubleConversion: explicit (double) casts in StatsCommand
- EffectivelyPrivate: drop public on nested-class methods (2 files)
- EscapedEntity: literal angle brackets inside {@code} in GuildDto
- InvalidParam: reword javadoc that referenced a non-parameter name

The explicit CheckSeverity.ERROR promotions from #446-#452 are retained as
defense-in-depth.

+++ src/main/java/io/taanielo/jmud/core/server/socket/StatsCommand.java
@@ -96,7 +96,7 @@ public class StatsCommand extends RegistrableCommand {
             Locale.ROOT,
             "Avg duration: %s ms | Max: %s ms | Overruns: %d (last %d ticks)",
             millis(summary.averageDurationNanos()),
-            millis(summary.maxDurationNanos()),
+            millis((double) summary.maxDurationNanos()),
             summary.overrunCount(),
             summary.windowTicks()
         ));
@@ -108,7 +108,7 @@ public class StatsCommand extends RegistrableCommand {
                 Locale.ROOT,
                 "Slowest Tickable: %s (%s ms / %d invocations)",
                 slowest,
-                millis(summary.slowestTickableTotalNanos()),
+                millis((double) summary.slowestTickableTotalNanos()),
                 summary.slowestTickableInvocations()
             ));
         }
